requirements.txt

Flask
selenium
webdriver-manager # To automatically manage chromedriver
python-dotenv